[zsh] polish flake update script
Better printing about whether mimeapps.list exists or not.

+++ modules/home/shell/zsh.nix
@@ -75,8 +75,13 @@
         }
 
         fsw() {
-          # echo "Deleting files that will be written"
-          rm -f "$HOME/.config/mimeapps.list"
+          if [[ ! -L "$HOME/.config/mimeapps.list" ]]; then
+            echo "Removing non-symlinked mimeapps.list"
+            rm -f "$HOME/.config/mimeapps.list"
+            echo "Exit status of rm is $?"
+          fi
+          echo -n "mimeapps.list should be a symlink or empty):"
+          echo "$(ls -l1 $HOME/.config/mimeapps.list)"
           sudo nixos-rebuild --flake "$HOME/workspace/areas/system-management/dotfiles/.#" switch "$@"
         }
       '';
